[yixuan-refactor] modularize
- moved unit_turn
- moved output sinks
