+++ prepare_childes.py
@@ -7,16 +7,15 @@ The default processing pipeline.
 """
 
 import argparse
-from contextlib import AbstractContextManager
 from functools import wraps
-import io
-import os
 from pathlib import Path
 import re
 from typing import Dict, List, Optional, Tuple
 
 import yaml
 
+from output import ChildesSink, ChildesSinkText
+from unit_turn import UnitTurn
 from utils import DataIntegrityError
 
 
@@ -60,82 +59,6 @@ class ChatConfig:
 
         except yaml.YAMLError as e:
             raise yaml.YAMLError(f'Error parsing YAML configuration: {str(e)}')
-
-
-class UnitTurn:
-    """Represent a unit of utterance with its surrounding environmental context.
-
-    This class encapsulates an utterance and its related environmental information
-    before, during, and after the utterance. It provides functionality to convert
-    the stored data into a consolidated list format.
-
-    Attributes:
-        head: A list representing the headers.
-        utt: A list representing the core utterance content.
-        sit: A list representing the environmental situation content.
-        env_bef: A list containing environmental events before the utterance.
-        env_dur: A list containing environmental events during the utterance.
-        env_aft: A list containing environmental events after the utterance.
-    """
-
-    def __init__(self) -> None:
-        """Initialize UnitUtterance with empty lists for utterance and environmental context.
-
-        The utterance and each environmental context (before, during, after)
-        are initialized as empty lists, ready to be populated with relevant data.
-        """
-        self.speaker = ''
-        self.head = []
-        self.utt = []
-        self.sit = []
-        self.env_bef = []
-        self.env_dur = []
-        self.env_aft = []
-
-    def update(self, content: str, attribute: str) -> None:
-        """Append content to the specified class attribute if it exists.
-
-        Args:
-            content: The content to be appended.
-            attribute: The name of the attribute to which the content will be appended.
-
-        Raises:
-            DataIntegrityError: If the specified attribute does not exist.
-        """
-        if hasattr(self, attribute):
-            attr = getattr(self, attribute)
-            if isinstance(attr, list):
-                attr.append(content)
-            elif isinstance(attr, str):
-                setattr(self, attribute, content)
-            else:
-                raise DataIntegrityError(
-                    f"Attribute '{attribute}' is not a list.",
-                    attribute
-                )
-        else:
-            raise DataIntegrityError(
-                f"Attribute '{attribute}' does not exist.",
-                attribute
-            )
-
-    def to_list(self) -> list:
-        """Convert the utterance and its environmental context into a single list.
-
-        This method consolidates the environmental context (before, during, after)
-        and the utterance itself into one sequential list.
-
-        Returns:
-            A list containing the combined data from env_bef, env_dur, utt, and env_aft.
-        """
-        output = []
-        output.extend(self.head)
-        output.extend(self.sit)
-        output.extend(self.env_bef)
-        output.extend(self.env_dur)
-        output.extend(self.utt)
-        output.extend(self.env_aft)
-        return output
 
 
 def process_basic(utterance: str, config: ChatConfig) -> str:
@@ -1140,44 +1063,6 @@ def split_sync_rel(input_line: str, config: ChatConfig) -> Dict[str, list]:
 
     return env_dict
 
-class ChildesSink(AbstractContextManager):
-    def write_turn(self, turn: UnitTurn) -> None:
-        ...
-
-class ChildesSinkText(ChildesSink):
-    """Text file implementation of ChildesSink that writes turns to a text file."""
-    
-    file_path: str
-    buffer: list[str]
-
-    def __init__(self, output_file: str):
-        """Initialize with the output file path.
-        
-        Args:
-            output_file: Path where the processed content will be written.
-        """
-        self.file_path = output_file
-        self.buffer = []
-
-    def __enter__(self):
-        """Open the file for writing and return self."""
-
-        os.makedirs(os.path.dirname(self.file_path), exist_ok=True)
-        return self
-    
-    def write_turn(self, turn: UnitTurn) -> None:
-        """Write a UnitTurn to the file."""
-        self.buffer.extend(turn.to_list())
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        try:
-            with open(self.file_path, "w", encoding="utf-8") as f:
-                f.writelines(self.buffer)
-        except Exception:
-            raise
-        else:
-            self.buffer.clear()
-
 def process_cha_file(input_file: str, output_file: str, config_path: str) -> None:
     """Read a .cha file and write cleaned conversations to output file.
 
@@ -1203,7 +1088,7 @@ def process_cha_file(input_file: str, output_file: str, config_path: str) -> Non
 
         # processed_lines = []
 
-        sink: ChildesSink = ChildesSinkText(output_file)
+        sink: ChildesSink = ChildesSinkText({"utterance"}, output_file=output_file)
 
         unit_turn = UnitTurn()
